Updated Ldap classes

+++ AmazonProdAdvtPlatform/src/main/java/com/amzn/model/nodes/ldapchilds/LdapChild.java
@@ -21,7 +21,7 @@ public class LdapChild extends AbstractLdapChild{
 
     @Override
     public void createAndLoadChildren() {
-	Map<String, Integer> leafChildrens=loadLeafs.loadRequiredChildren(this.parentCategory);
+	Map<String, Long> leafChildrens=loadLeafs.loadRequiredChildren(this.parentCategory);
 	System.out.println(leafChildrens);
     }
